Updated names to reflect blog instead of "dbtest"; moved & renamed location for db model file(s)
