+++ views/blog.ejs
@@ -1,0 +1,28 @@
+<!-- views/dbtest.ejs -->
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <%- include('partials/head') %>
+</head>
+<body class="container">
+
+<header>
+    <%- include('partials/header') %>
+</header>
+
+<main>
+    <div class="jumbotron">
+        <h1><%=title%></h1>
+<% for (var i = 0; i < posts.length; i++) { %>
+        <h3><%= posts[i].title %></h3>
+	<h5><%= posts[i].body %></h5>
+<% } %>
+    </div>
+</main>
+
+<footer>
+    <%- include('partials/footer') %>
+</footer>
+
+</body>
+</html>
