+++ views/partials/header.ejs
@@ -68,7 +68,7 @@
                 </li>
             </ul>
             <ul class="nav navbar-nav navbar-right">
-                <li><a href="/dbtest"><span class="glyphicon glyphicon-book"></span> My Blog</a></li>
+                <li><a href="/blog"><span class="glyphicon glyphicon-book"></span> My Blog</a></li>
             </ul>
         </div>
     </div>
